+++ R/ui.R
@@ -8,7 +8,9 @@
         shiny::actionButton("return_button", "Return", icon = shiny::icon("door-open"),
                             style = "position: absolute; top: 8px; right: 8px; z-index:10000;"),
         shiny::tagAppendAttributes(
-            shiny::textOutput("datetime_range"), style="font-weight: bold; color: white; position: absolute; top: 20px; right: 120px; z-index:10000;"),
+            shiny::textOutput("selected_item_text"), style="font-weight: bold; color: white; position: absolute; top: 10px; right: 120px; z-index:10000;"),
+        shiny::tagAppendAttributes(
+            shiny::textOutput("datetime_range_text"), style="font-weight: bold; color: white; position: absolute; top: 30px; right: 120px; z-index:10000;"),
         shinyjs::useShinyjs(),
         shiny::navbarPage("myClimGui",
                           .ui_plot_tab(data, data_loggers),